backend/pyproject.toml
[project]
name = "invoice-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.128.0",
    "httpx[socks]>=0.28.1",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "zhipuai>=2.1.5.20250825",
]
